+++ crwn-clothing/src/store/store.js
@@ -1,20 +1,11 @@
 import {compose, legacy_createStore as createStore , applyMiddleware} from 'redux';
-// import logger from 'redux-logger';
+import logger from 'redux-logger';
 import { persistReducer, persistStore } from "redux-persist";
 import storage from 'redux-persist/lib/storage';
 import { rootReducer } from './root-reducer';
 
-const loggerMiddleWare = (store)=>(next)=>(action)=>{
-    if(!action.type){
-        return next(action);
-    }
-    console.log('type', action.type);
-    console.log('payload', action.payload);
-    console.log('state', store.getState());
-    next(action);
-    console.log('state', store.getState());
-}
-const middleWares = [loggerMiddleWare];
+
+const middleWares = [process.env.NODE_ENV ==="development" &&logger].filter(Boolean);
 
 const composedEnhancers = compose(applyMiddleware(...middleWares));
